+++ deepfellow/server/__init__.py
@@ -20,6 +20,7 @@ from .organization import app as organization_app
 from .password_reset import app as reset_password_app
 from .project import app as project_app
 from .start import app as start_app
+from .status import app as status_app
 from .stop import app as stop_app
 from .update import app as update_app
 
@@ -31,6 +32,7 @@ app.add_typer(create_admin_app)
 app.add_typer(opentelemetry_app)
 app.add_typer(reset_password_app)
 app.add_typer(start_app)
+app.add_typer(status_app)
 app.add_typer(stop_app)
 app.add_typer(update_app)
 app.add_typer(logs_app)