+++ deepfellow/server/status.py
@@ -1,0 +1,31 @@
+# DeepFellow Software Framework.
+# Copyright © 2026 Simplito sp. z o.o.
+#
+# This file is part of the DeepFellow Software Framework (https://deepfellow.ai).
+# This software is Licensed under the DeepFellow Free License.
+#
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Start server typer command."""
+
+from pathlib import Path
+
+import typer
+
+from deepfellow.common.docker import print_docker_status
+from deepfellow.common.echo import echo
+from deepfellow.common.install import assert_docker
+from deepfellow.server.utils.options import directory_option
+
+app = typer.Typer()
+
+
+@app.command()
+def status(
+    directory: Path = directory_option(exists=True),
+) -> None:
+    """Show DeepFellow Server status."""
+    assert_docker()
+    echo.debug("Showing DeepFellow Server status")
+    print_docker_status(directory, "server")